requirements.txt
MarkupSafe==2.1.3
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
jinja2==3.1.2
orjson==3.9.10
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
cffi==1.16.0
fastapi_login==1.9.1
outcome==1.3.0.post0
passlib==1.7.4
pycparser==2.21
pyjwt==2.8.0
sortedcontainers==2.4.0
trio==0.21.0
Mako==1.2.4
alembic==1.12.1
greenlet==3.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
cryptography==41.0.5
email-validator==2.0.0.post2
fastapi_users==12.1.2
makefun==1.15.1
fastapi_users_db_sqlalchemy==6.0.1
async-timeout==4.0.3
asyncpg==0.29.0
charset-normalizer==3.3.2
requests==2.31.0
urllib3==2.1.0
Pillow==10.1.0



